WPC for demonstration at ITU, BEXCO 20170925-0928
ITU expo is going to held in Busan and GCR is allowed to participate as a selected 20 of K-Global 300 companies. We are going to demonstrate the Walnut Programmable Car and the Self-Driving Car using Arduino as well as Ultrasound Measurement System for Height and Distance. // WPC_ITUdemo.ino is a main code accompanied with RealyCar.ino